requirements.txt
Django~=5.1
djangorestframework~=3.15
djangorestframework-simplejwt~=5.2
python-dotenv~=1.0
pip
